refactor(snap-keyring): split helpers show{Success,Error}/getSnapName into files
