+++ app/scripts/lib/snap-keyring/utils/getSnapName.ts
@@ -1,0 +1,14 @@
+import { getSnapName as getSnapNameFromSubjectMetadata } from '../../../../../ui/helpers/utils/util';
+import { SnapKeyringBuilderMessenger } from '../types';
+
+export const getSnapName = (
+  controllerMessenger: SnapKeyringBuilderMessenger,
+  snapId: string,
+): string => {
+  const subjectMetadata = controllerMessenger.call(
+    'SubjectMetadataController:getSubjectMetadata',
+    snapId,
+  );
+
+  return getSnapNameFromSubjectMetadata(snapId, subjectMetadata);
+};
